- moved eepriom write pin hardware functions to system.c - rename tof_write_protect to tof_enable_pin - change BUFFER to be uint8_t instead of uint32_t. - cleaning up

+++ stm32-modules/include/flex-stacker/flex-stacker/gcodes.hpp
@@ -428,7 +428,7 @@ struct GetTOFRegister {
 struct SetTOFRegister {
     TOFSensorID sensor_id;
     uint8_t reg;
-    uint32_t data;
+    uint8_t data;
 
     using ParseResult = std::optional<SetTOFRegister>;
     static constexpr auto prefix = std::array{'M', '2', '2', '3', ' '};
